Removed color_shift

+++ code/User_UI/user_ui.cpp
@@ -312,10 +312,16 @@ inline void showSEEK(SETTINGS_UI_MENU_LIST element)  // Prikaze element v seeku 
         case SU_STATE_STRIP_SELECTION:
             if (element.index == SU_MENU_STRIP_ANIMATION_STRIP_OFF)
             {
-                deleteTASK(&m_audio_system.handle_active_strip_mode);
-                m_audio_system.set_strip_brightness(255);
-                m_audio_system.color_shift(COLOR_RED, AUVS_CFG_NORMAL_ANIMATION_TIME_MS);
+                if (m_audio_system.strip.curr_color_index != COLOR_RED)
+                {
+                    deleteTASK(&m_audio_system.handle_active_strip_mode);
+                    brightDOWN(AUVS_CFG_SLOW_ANIMATION_TIME_MS);
+                    m_audio_system.set_strip_brightness(255);
+                    m_audio_system.set_strip_color(COLOR_RED);
+                    brightUP(AUVS_CFG_SLOW_ANIMATION_TIME_MS);
+                }
             }
+
             else if (m_audio_system.handle_active_strip_mode == NULL)
             {
                 m_audio_system.create_animation(element.index, COLOR_WHITE);
